Regularization paths for regression models with grouped covariates

This package fits regularization paths for linear, logistic, and Cox regression models with grouped penalties, such as the group lasso, group MCP, group SCAD, group exponential lasso, and group bridge. The algorithms are based on the idea of either locally approximated coordinate descent or group descent, depending on the penalty. All of the algorithms (with the exception of group bridge) are stable and fast.

Details

Given a design matrix X in which the features consist of non-overlapping groups and vector of responses y, grpreg solves the regularization path for a variety of penalties. The package also provides methods for plotting and cross-validation.

See the "Getting started" vignette for a brief overview of how the package works.

The following penalties are available:

grLasso: Group lasso (Yuan and Lin, 2006)

grMCP: Group MCP; like the group lasso, but with an MCP penalty on the norm of each group

grSCAD: Group SCAD; like the group lasso, but with a SCAD penalty on the norm of each group

cMCP: A hierarchical penalty which places an outer MCP penalty on a sum of inner MCP penalties for each group (Breheny & Huang, 2009)

gel: Group exponential lasso (Breheny, 2015)

gBridge: A penalty which places a bridge penalty on the L1-norm of each group (Huang et al., 2009)

The cMCP, gel, and gBridge penalties carry out bi-level selection, meaning that they carry out variable selection at the group level and at the level of individual covariates (i.e., they select important groups as well as important members of those groups). The grLasso, grMCP, and grSCAD penalties carry out group selection, meaning that within a group, coefficients will either all be zero or all nonzero. A variety of supporting methods for selecting lambda and plotting the paths are provided also.

References

Yuan M and Lin Y. (2006) Model selection and estimation in regression with grouped variables. Journal of the Royal Statistical Society Series B, 68: 49-67. doi:10.1111/j.1467-9868.2005.00532.x

Huang J, Ma S, Xie H, and Zhang C. (2009) A group bridge approach for variable selection. Biometrika, 96: 339-355. doi:10.1093/biomet/asp020

Breheny P and Huang J. (2009) Penalized methods for bi-level variable selection. Statistics and its interface, 2: 369-380. doi:10.4310/sii.2009.v2.n3.a10

Huang J, Breheny P, and Ma S. (2012). A selective review of group selection in high dimensional models. Statistical Science, 27: 481-499. doi:10.1214/12-sts392

Breheny P and Huang J. (2015) Group descent algorithms for nonconvex penalized linear and logistic regression models with grouped predictors. Statistics and Computing, 25: 173-187. doi:10.1007/s11222-013-9424-2

Breheny P. (2015) The group exponential lasso for bi-level variable selection. Biometrics, 71: 731-740. doi:10.1111/biom.12300

Author

Patrick Breheny

Examples

if (FALSE) {
vignette("getting-started", "grpreg")
}