Summarizing inferences based on cross-validation

Summary method for cv.grpreg or cv.grpsurv objects

# S3 method for cv.grpreg
summary(object, ...)
# S3 method for summary.cv.grpreg
print(x, digits, ...)

Arguments

object

A "cv.grpreg" object.

x

A "summary.cv.grpreg" object.

digits

Number of digits past the decimal point to print out. Can be a vector specifying different display digits for each of the five non-integer printed values.

...

Further arguments passed to or from other methods.

Value

summary(cvfit) produces an object with S3 class

"summary.cv.grpreg". The class has its own print method and contains the following list elements:

penalty

The penalty used by grpreg/grpsurv.

model

The type of model: "linear", "logistic", "Poisson", "Cox", etc.

n

Number of observations

p

Number of regression coefficients (not including the intercept).

min

The index of lambda with the smallest cross-validation error.

lambda

The sequence of lambda values used by cv.grpreg/cv.grpsurv.

cve

Cross-validation error (deviance).

r.squared

Proportion of variance explained by the model, as estimated by cross-validation.

snr

Signal to noise ratio, as estimated by cross-validation.

sigma

For linear regression models, the scale parameter estimate.

pe

For logistic regression models, the prediction error (misclassification error).

Author

Patrick Breheny

See also

Examples

# Birthweight data
data(Birthwt)
X <- Birthwt$X
group <- Birthwt$group

# Linear regression
y <- Birthwt$bwt
cvfit <- cv.grpreg(X, y, group)
summary(cvfit)
#> grLasso-penalized linear regression with n=189, p=16
#> At minimum cross-validation error (lambda=0.0184):
#> -------------------------------------------------
#>   Nonzero coefficients: 16
#>   Nonzero groups: 8
#>   Cross-validation error of 0.43
#>   Maximum R-squared: 0.19
#>   Maximum signal-to-noise ratio: 0.23
#>   Scale estimate (sigma) at lambda.min: 0.656

# Logistic regression
y <- Birthwt$low
cvfit <- cv.grpreg(X, y, group, family="binomial")
summary(cvfit)
#> grLasso-penalized logistic regression with n=189, p=16
#> At minimum cross-validation error (lambda=0.0198):
#> -------------------------------------------------
#>   Nonzero coefficients: 16
#>   Nonzero groups: 8
#>   Cross-validation error of 1.17
#>   Maximum R-squared: 0.07
#>   Maximum signal-to-noise ratio: 0.06
#>   Prediction error at lambda.min: 0.323

# Cox regression
data(Lung)
cvfit <- with(Lung, cv.grpsurv(X, y, group))
summary(cvfit)
#> grLasso-penalized Cox regression with n=137, p=14
#> At minimum cross-validation error (lambda=0.1063):
#> -------------------------------------------------
#>   Nonzero coefficients: 7
#>   Nonzero groups: 2
#>   Cross-validation error of 7.61
#>   Maximum R-squared: 0.25
#>   Maximum signal-to-noise ratio: 0.04